requirements.txt
google-generativeai
gradio
python-dotenv
flask
gunicorn
